% Auto-generated: do not edit by hand
\name{''Parallel}

\alias{''Parallel}

\title{Parallel component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''Parallel(id=NULL, color_encode_columns=NULL, data=NULL, height=NULL,
line=NULL, selectedPath=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{color_encode_columns}{List of characters. }

\item{data}{List of named lists. }

\item{height}{Numeric. }

\item{line}{Named list. }

\item{selectedPath}{List of characters. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

