% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{pcCenter}
\alias{pcCenter}
\title{Center \code{parcoords} horizontally based on column/variable through \code{parcoordsProxy}}
\usage{
pcCenter(pc = NULL, dim = NULL)
}
\arguments{
\item{pc}{\code{parcoordsProxy}}

\item{dim}{\code{string} column/variable to center.}
}
\value{
\code{parcoords_proxy}
}
\description{
Center \code{parcoords} horizontally based on column/variable through \code{parcoordsProxy}
}
