% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{pcFilter}
\alias{pcFilter}
\title{Filter \code{parcoords} through \code{parcoordsProxy}}
\usage{
pcFilter(pc = NULL, filters = NULL)
}
\arguments{
\item{pc}{\code{parcoordsProxy}}

\item{filters}{\code{list} of filters to apply to the parcoords proxy.  Please see
\href{https://github.com/deitch/searchjs}{search.js} for example queries as filters.}
}
\value{
\code{parcoords_proxy}
}
\description{
Filter \code{parcoords} through \code{parcoordsProxy}
}
