% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{pcSnapshot}
\alias{pcSnapshot}
\title{Download image of \code{parcoords} through \code{parcoordsProxy}}
\usage{
pcSnapshot(pc = NULL)
}
\arguments{
\item{pc}{\code{parcoordsProxy}}
}
\value{
\code{parcoords_proxy}
}
\description{
Download image of \code{parcoords} through \code{parcoordsProxy}
}
