% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{pcUnhide}
\alias{pcUnhide}
\title{Unhide \code{parcoords} columns through \code{parcoordsProxy}}
\usage{
pcUnhide(pc = NULL, dim = NULL)
}
\arguments{
\item{pc}{\code{parcoordsProxy}}

\item{dim}{\code{string} column(s) to hide.}
}
\value{
\code{parcoords_proxy}
}
\description{
Unhide \code{parcoords} columns through \code{parcoordsProxy}
}
