% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoords.R
\name{renderParcoords}
\alias{renderParcoords}
\title{Widget render function for use in Shiny}
\usage{
renderParcoords(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
\seealso{
\code{\link{parcoordsOutput}}
}

