% Auto-generated: do not edit by hand
\name{''Parallel}

\alias{''Parallel}

\title{Parallel component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''Parallel(id=NULL, color_encode_column=NULL, data=NULL, do_color=NULL,
height=NULL, line=NULL, margin=NULL, ordinal_scale=NULL,
width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{color_encode_column}{Character. }

\item{data}{List of named lists. }

\item{do_color}{Logical. }

\item{height}{Numeric. }

\item{line}{Named list. }

\item{margin}{Lists containing elements 'top', 'right', 'bottom', 'left'.
those elements have the following types:
  - top (numeric; optional)
  - right (numeric; optional)
  - bottom (numeric; optional)
  - left (numeric; optional). }

\item{ordinal_scale}{Logical. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

